feat(DTFS2-7393): move clone & maker tests into ff disabled directory
